backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.1.8
cloudpickle==3.1.2
colorama==0.4.6
exceptiongroup==1.3.1
fairlearn==0.13.0
fastapi==0.128.8
h11==0.16.0
idna==3.11
joblib==1.5.3
llvmlite==0.43.0
narwhals==2.19.0
numba==0.60.0
numpy==2.0.2
packaging==26.1
pandas==2.3.3
pydantic==2.13.2
pydantic_core==2.46.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
scikit-learn==1.6.1
scipy==1.13.1
shap==0.49.1
six==1.17.0
slicer==0.0.8
starlette==0.49.3
threadpoolctl==3.6.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1

# ── newly added ──
google-generativeai>=0.8.0
firebase-admin>=6.5.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
uvicorn==0.39.0
